Refactor ResourcesAdd to avoid repetition
